Add files via upload
The ATM is a self-service banking machine that provides customers with 24/7 access to a wide range of financial services. With its easy-to-use interface, the ATM allows users to withdraw cash, check account balances, transfer funds between accounts, and even deposit money. It securely verifies your identity using a personal identification number (PIN) to ensure the safety of your transactions. Whether you're in need of cash, making quick transfers, or checking your balance, the ATM offers a fast, convenient, and reliable way to manage your finances without the need to visit a bank branch. With locations worldwide, the ATM is designed to offer a seamless, efficient, and secure banking experience anytime, anywhere.